(SERVER-2265) Refactor: move puppet config
