+++ lib/puppetserver/ca/action/create.rb
@@ -1,6 +1,6 @@
 require 'puppetserver/ca/utils/cli_parsing'
 require 'puppetserver/ca/host'
-require 'puppetserver/ca/puppet_config'
+require 'puppetserver/ca/config/puppet'
 require 'puppetserver/ca/utils/file_system'
 require 'puppetserver/ca/utils/http_client'
 require 'puppetserver/utils/signing_digest'
