+++ lib/puppetserver/ca/action/revoke.rb
@@ -1,7 +1,7 @@
 require 'puppetserver/ca/utils/cli_parsing'
 require 'puppetserver/ca/utils/http_client'
 require 'puppetserver/ca/utils/file_system'
-require 'puppetserver/ca/puppet_config'
+require 'puppetserver/ca/config/puppet'
 
 require 'optparse'
 require 'json'
