+++ lib/puppetserver/ca/config/puppet.rb
